src/python/requirements.txt
# Python dependencies
requests==2.31.0
flask==2.3.3
flask-cors==4.0.0
python-dotenv==1.0.0
openai==1.8.0
# Add your Python package requirements here
# Example:
# pandas==2.1.0 